requirements.txt
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.12.1
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
